chore: inline getBoundingSize
