+++ lib/core/base/create-frame-context.js
@@ -1,5 +1,3 @@
-import { getBoundingSize } from '../utils';
-
 export default function createFrameContext(frame, { focusable, page }) {
   return {
     node: frame,
@@ -19,4 +17,16 @@ function frameFocusable(frame) {
   }
   const int = parseInt(tabIndex, 10)
   return isNaN(int) || int >= 0;
+}
+
+function getBoundingSize(domNode) {
+  let width = parseInt(domNode.getAttribute('width'), 10);
+  let height = parseInt(domNode.getAttribute('height'), 10);
+
+  if (isNaN(width) || isNaN(height)) {
+    const rect = domNode.getBoundingClientRect();
+    width = isNaN(width) ? rect.width : width;
+    height = isNaN(height) ? rect.height : height;
+  }
+  return { width, height }
 }
