+++ lib/core/utils/index.js
@@ -27,7 +27,6 @@ export { default as findBy } from './find-by';
 export { default as getFlattenedTree } from './get-flattened-tree';
 export { default as getAllChecks } from './get-all-checks';
 export { default as getBaseLang } from './get-base-lang';
-export { default as getBoundingSize } from './get-bounding-size';
 export { default as getCheckMessage } from './get-check-message';
 export { default as getCheckOption } from './get-check-option';
 export { default as getFrameContexts } from './get-frame-contexts';
